move database config
